Ignore error and log in renderGeoloniaMap(); re-throw init errors.
When an error is thrown in renderGeoloniaMap(), subsequent maps will fail
to render and show the error dialog.

Additionally, the try/catch was shortened to include only the new map (super(options))
and re-throws the initial error object, because at that point the map is useless and
should not be used normally. If the error was not rethrown, code using the map
wouldn't know if the map was valid or not.